+++ atlasdb-client/src/main/java/com/palantir/atlasdb/table/description/render/StreamStoreRenderer.java
@@ -39,6 +39,7 @@ import javax.annotation.Generated;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import com.google.common.base.Functions;
 import com.google.common.base.Preconditions;
 import com.google.common.collect.ArrayListMultimap;
 import com.google.common.collect.Collections2;
@@ -50,6 +51,7 @@ import com.google.common.collect.Maps;
 import com.google.common.collect.Multimap;
 import com.google.common.collect.Multimaps;
 import com.google.common.collect.Sets;
+import com.google.common.io.ByteStreams;
 import com.google.common.io.CountingInputStream;
 import com.google.common.primitives.Ints;
 import com.google.protobuf.ByteString;
@@ -66,12 +68,16 @@ import com.palantir.atlasdb.transaction.api.TransactionManager;
 import com.palantir.atlasdb.transaction.api.TransactionTask;
 import com.palantir.atlasdb.transaction.impl.TxTask;
 import com.palantir.common.base.Throwables;
+import com.palantir.common.compression.LZ4CompressingInputStream;
 import com.palantir.common.io.ConcatenatedInputStream;
 import com.palantir.util.AssertUtils;
 import com.palantir.util.ByteArrayIOStream;
+import com.palantir.util.Pair;
 import com.palantir.util.crypto.Sha256Hash;
 import com.palantir.util.file.DeleteOnCloseFileInputStream;
 import com.palantir.util.file.TempFileUtils;
+
+import net.jpountz.lz4.LZ4BlockInputStream;
 
 public class StreamStoreRenderer {
     private final String name;
@@ -79,13 +85,15 @@ public class StreamStoreRenderer {
     private final String packageName;
     private final String schemaName;
     private final int inMemoryThreshold;
+    private final boolean clientSideCompression;
 
-    public StreamStoreRenderer(String name, ValueType streamIdType, String packageName, String schemaName, int inMemoryThreshold) {
+    public StreamStoreRenderer(String name, ValueType streamIdType, String packageName, String schemaName, int inMemoryThreshold, boolean clientSideCompression) {
         this.name = name;
         this.streamIdType = streamIdType;
         this.packageName = packageName;
         this.schemaName = schemaName;
         this.inMemoryThreshold = inMemoryThreshold;
+        this.clientSideCompression = clientSideCompression;
     }
 
     public String getPackageName() {
@@ -156,6 +164,16 @@ public class StreamStoreRenderer {
                     line();
                     getBlock();
                     line();
+                    if (clientSideCompression) {
+                        storeBlocksAndGetFinalMetadata();
+                        line();
+                        loadStreamWithCompression();
+                        line();
+                        loadStreamsWithCompression();
+                        line();
+                        tryWriteStreamToFile();
+                        line();
+                    }
                     getMetadata();
                     line();
                     lookupStreamIdsByHash();
@@ -515,6 +533,52 @@ public class StreamStoreRenderer {
                     line("index.delete(toDelete);");
                 } line("}");
             }
+
+            private void storeBlocksAndGetFinalMetadata() {
+                line("@Override");
+                line("protected StreamMetadata storeBlocksAndGetFinalMetadata(Transaction t, long id, InputStream stream) {"); {
+                    line("//Hash the data before compressing it");
+                    line("MessageDigest digest = Sha256Hash.getMessageDigest();");
+                    line("try (InputStream hashingStream = new DigestInputStream(stream, digest);");
+                    line("        InputStream compressingStream = new LZ4CompressingInputStream(hashingStream)) {"); {
+                        line("StreamMetadata metadata = storeBlocksAndGetHashlessMetadata(t, id, compressingStream);");
+                        line("return StreamMetadata.newBuilder(metadata)");
+                        line("        .setHash(ByteString.copyFrom(digest.digest()))");
+                        line("        .build();");
+                    } line("} catch (IOException e) {"); {
+                        line("throw new RuntimeException(e);");
+                    } line("}");
+                } line("}");
+            }
+
+            private void loadStreamWithCompression() {
+                line("@Override");
+                line("public InputStream loadStream(Transaction t, final ", StreamId, " id) {"); {
+                    line("return new LZ4BlockInputStream(super.loadStream(t, id));");
+                } line("}");
+            }
+
+            private void loadStreamsWithCompression() {
+                line("@Override");
+                line("public Map<", StreamId, ", InputStream> loadStreams(Transaction t, Set<", StreamId, "> ids) {"); {
+                    line("Map<", StreamId, ", InputStream> compressedStreams = super.loadStreams(t, ids);");
+                    line("return Maps.transformValues(compressedStreams, stream -> {"); {
+                        line("return new LZ4BlockInputStream(stream);");
+                    } line("});");
+                } line("}");
+            }
+
+            private void tryWriteStreamToFile() {
+                line("@Override");
+                line("protected void tryWriteStreamToFile(Transaction transaction, ", StreamId, " id, StreamMetadata metadata, FileOutputStream fos) throws IOException {"); {
+                    line("try (InputStream blockStream = makeStream(id, metadata);");
+                    line("        InputStream decompressingStream = new LZ4BlockInputStream(blockStream);");
+                    line("        OutputStream fileStream = fos;) {"); {
+                        line("ByteStreams.copy(decompressingStream, fileStream);");
+                    } line("}");
+                } line("}");
+            }
+
         }.render();
     }
 
@@ -712,5 +776,10 @@ public class StreamStoreRenderer {
         List.class,
         CheckForNull.class,
         Generated.class,
+        LZ4CompressingInputStream.class,
+        LZ4BlockInputStream.class,
+        Pair.class,
+        Functions.class,
+        ByteStreams.class,
     };
 }